separate map file, code formatting, add persave config
